Fix FelizSvelte
