+++ Fable.Svelte.sln
@@ -17,6 +17,12 @@ Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "App", "samples\App\src\App.
 EndProject
 Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.SveltePlugins", "src\Fable.SveltePlugins\Fable.SveltePlugins.fsproj", "{660DA086-4063-40CD-9AA3-C800BC858DB1}"
 EndProject
+Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.ReactStore", "src\Fable.ReactStore\Fable.ReactStore.fsproj", "{15F0C697-49CE-4706-A84F-479F22CED14B}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "FelizSvelte", "FelizSvelte", "{C03D6FC1-1ABE-43BD-907B-9F68DF2F1267}"
+EndProject
+Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "FelizSvelte", "samples\FelizSvelte\src\FelizSvelte.fsproj", "{37724441-3E93-4970-AC6B-F8A1E395CBCA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -78,6 +84,30 @@ Global
 		{660DA086-4063-40CD-9AA3-C800BC858DB1}.Release|x64.Build.0 = Release|Any CPU
 		{660DA086-4063-40CD-9AA3-C800BC858DB1}.Release|x86.ActiveCfg = Release|Any CPU
 		{660DA086-4063-40CD-9AA3-C800BC858DB1}.Release|x86.Build.0 = Release|Any CPU
+		{15F0C697-49CE-4706-A84F-479F22CED14B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{15F0C697-49CE-4706-A84F-479F22CED14B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{15F0C697-49CE-4706-A84F-479F22CED14B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{15F0C697-49CE-4706-A84F-479F22CED14B}.Debug|x64.Build.0 = Debug|Any CPU
+		{15F0C697-49CE-4706-A84F-479F22CED14B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{15F0C697-49CE-4706-A84F-479F22CED14B}.Debug|x86.Build.0 = Debug|Any CPU
+		{15F0C697-49CE-4706-A84F-479F22CED14B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{15F0C697-49CE-4706-A84F-479F22CED14B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{15F0C697-49CE-4706-A84F-479F22CED14B}.Release|x64.ActiveCfg = Release|Any CPU
+		{15F0C697-49CE-4706-A84F-479F22CED14B}.Release|x64.Build.0 = Release|Any CPU
+		{15F0C697-49CE-4706-A84F-479F22CED14B}.Release|x86.ActiveCfg = Release|Any CPU
+		{15F0C697-49CE-4706-A84F-479F22CED14B}.Release|x86.Build.0 = Release|Any CPU
+		{37724441-3E93-4970-AC6B-F8A1E395CBCA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{37724441-3E93-4970-AC6B-F8A1E395CBCA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{37724441-3E93-4970-AC6B-F8A1E395CBCA}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{37724441-3E93-4970-AC6B-F8A1E395CBCA}.Debug|x64.Build.0 = Debug|Any CPU
+		{37724441-3E93-4970-AC6B-F8A1E395CBCA}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{37724441-3E93-4970-AC6B-F8A1E395CBCA}.Debug|x86.Build.0 = Debug|Any CPU
+		{37724441-3E93-4970-AC6B-F8A1E395CBCA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{37724441-3E93-4970-AC6B-F8A1E395CBCA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{37724441-3E93-4970-AC6B-F8A1E395CBCA}.Release|x64.ActiveCfg = Release|Any CPU
+		{37724441-3E93-4970-AC6B-F8A1E395CBCA}.Release|x64.Build.0 = Release|Any CPU
+		{37724441-3E93-4970-AC6B-F8A1E395CBCA}.Release|x86.ActiveCfg = Release|Any CPU
+		{37724441-3E93-4970-AC6B-F8A1E395CBCA}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{517E93EC-2D1D-4FBD-9763-762F0868B34C} = {C79CCF59-F99B-47B9-AFBC-04C5CFE317F1}
@@ -85,5 +115,8 @@ Global
 		{1AB18684-BA9F-4DF2-9BA9-4B43A2E4B779} = {75848987-1EA4-4C34-845A-9AB14B76CDD1}
 		{3DED0940-98BC-4F0B-BA3F-7BF1A84B55F3} = {1AB18684-BA9F-4DF2-9BA9-4B43A2E4B779}
 		{660DA086-4063-40CD-9AA3-C800BC858DB1} = {C79CCF59-F99B-47B9-AFBC-04C5CFE317F1}
+		{15F0C697-49CE-4706-A84F-479F22CED14B} = {C79CCF59-F99B-47B9-AFBC-04C5CFE317F1}
+		{C03D6FC1-1ABE-43BD-907B-9F68DF2F1267} = {75848987-1EA4-4C34-845A-9AB14B76CDD1}
+		{37724441-3E93-4970-AC6B-F8A1E395CBCA} = {C03D6FC1-1ABE-43BD-907B-9F68DF2F1267}
 	EndGlobalSection
 EndGlobal
