+++ samples/FelizSvelte/.config/dotnet-tools.json
@@ -1,0 +1,12 @@
+{
+  "version": 1,
+  "isRoot": true,
+  "tools": {
+    "fable": {
+      "version": "3.1.1",
+      "commands": [
+        "fable"
+      ]
+    }
+  }
+}
