+++ samples/FelizSvelte/src/FelizSvelte.fsproj
@@ -5,11 +5,10 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="Feliz.Store.fs" />
     <Compile Include="TimeFliesElmish.fs" />
     <Compile Include="Util.fs" />
     <Compile Include="Dragging.fs" />
-    <Compile Include="TodoMvc.fs" />
+    <Compile Include="TodoMVC.fs" />
     <Compile Include="App.fs" />
     <Compile Include="Main.fs" />
   </ItemGroup>
@@ -20,7 +19,8 @@
     <PackageReference Include="Feliz.Router" Version="3.5.0" />
     <PackageReference Include="Feliz.Svelte" Version="0.4.0" />
     <PackageReference Update="FSharp.Core" Version="4.7.2" />
-    <ProjectReference Include="..\..\..\src\Fable.Svelte\Fable.Svelte.fsproj" />
+    <ProjectReference Include="..\..\..\src\Fable.SvelteStore\Fable.SvelteStore.fsproj" />
+    <ProjectReference Include="..\..\..\src\Fable.ReactStore\Fable.ReactStore.fsproj" />
   </ItemGroup>
 
 </Project>
